refactor for empty role

+++ features/roles/AccessControlService.ts
@@ -1,4 +1,4 @@
-import { AccessControl } from 'accesscontrol';
+import { AccessControl, Permission } from 'accesscontrol';
 import { isEmpty } from 'lodash';
 
 export type Role = {
@@ -11,6 +11,7 @@ export type Role = {
 export default class AccessControlService {
   public ac;
   public role;
+  private falsePermission: Permission = {granted: false} as Permission;
 
   constructor(role: Role) {
     this.ac = new AccessControl();
@@ -21,35 +22,49 @@ export default class AccessControlService {
     this.setPermissionsForRecord();
   }
 
-  get roleName() {
+  get roleName(): string {
+    if(!this.role) return "";
+
     return this.role?.name;
   }
 
-  public hasRole(roleName: string) {
+  public hasRole(roleName: string): boolean {
+    if(!this.role) return false;
+
     return this.roleName === roleName;
   }
 
-  public createAny(record: string) {
+  public createAny(record: string): Permission {
+    if(!this.role) return this.falsePermission;
+
     return this.ac.can(this.roleName).createAny(record);
   }
 
-  public readAny(record: string) {
+  public readAny(record: string): Permission {
+    if(!this.role) return this.falsePermission;
+
     return this.ac.can(this.roleName).readAny(record);
   }
 
-  public updateAny(record: string) {
+  public updateAny(record: string): Permission {
+    if(!this.role) return this.falsePermission;
+
     return this.ac.can(this.roleName).updateAny(record);
   }
 
-  public deleteAny(record: string) {
+  public deleteAny(record: string): Permission {
+    if(!this.role) return this.falsePermission;
+
     return this.ac.can(this.roleName).deleteAny(record);
   }
 
-  private getRoleAbilityLogic(ability: string) {
+  private getRoleAbilityLogic(ability: string): boolean | undefined {
+    if(!this.role) return false;
+
     return isEmpty(this.role.options) || (this.role.options.abilities && this.role.options.abilities.includes(ability));
   }
 
-  private setPermissionsForRecord() {
+  private setPermissionsForRecord(): void {
     if(this.getRoleAbilityLogic("can_create")) this.ac.grant(this.roleName).createAny("record");
     if(this.getRoleAbilityLogic("can_read")) this.ac.grant(this.roleName).readAny("record");
     if(this.getRoleAbilityLogic("can_update")) this.ac.grant(this.roleName).updateAny("record");
